Small bug in one test
